+++ offer-manager-app/src/main/groovy/life/qbic/portal/offermanager/components/Resettable.groovy
@@ -1,0 +1,17 @@
+package life.qbic.portal.offermanager.components
+
+/**
+ * Reset graphical input elements
+ *
+ * A class that inherits the Resettable interface indicates, that
+ * it supports full reset of graphical input field elements.
+ *
+ * @since 1.0.0
+ */
+interface Resettable {
+
+    /**
+     * Resets all input elements of an graphical user interface
+     */
+    void reset()
+}