+++ offer-manager-app/src/main/groovy/life/qbic/portal/offermanager/components/offer/create/CreateOfferPresenter.groovy
@@ -31,6 +31,7 @@ class CreateOfferPresenter implements CreateOfferOutput, FetchOfferOutput{
                 "\'${createdOffer.projectTitle}\' successfully")
 
         this.offerService.addToResource(createdOffer)
+        this.createOfferViewModel.setOfferCreatedSuccessfully(true)
     }
 
     @Override